refactor: move xpath tests to correct package
