+++ src/test/kotlin/com/asyncapi/plugin/idea/_core/xpath/YamlFileXPathTest.kt
@@ -1,16 +1,16 @@
-package com.asyncapi.plugin.idea._core
+package com.asyncapi.plugin.idea._core.xpath
 
-import com.asyncapi.plugin.idea._core.xpath.JsonFileXPath
-import com.intellij.json.psi.JsonFile
 import com.intellij.lang.Language
 import com.intellij.psi.PsiFileFactory
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
 import junit.framework.TestCase
+import org.jetbrains.yaml.YAMLLanguage
+import org.jetbrains.yaml.psi.YAMLFile
 
 /**
  * @author Pavel Bodiachevskii
  */
-class JsonFileXPathTest: BasePlatformTestCase() {
+class YamlFileXPathTest: BasePlatformTestCase() {
 
     fun `test - 2_0_0`() {
         test("asyncapi", "2.0.0")
@@ -42,18 +42,18 @@ class JsonFileXPathTest: BasePlatformTestCase() {
 
     fun test(asyncapi: String, version: String) {
         val psiFileFactory = PsiFileFactory.getInstance(project)
-        val asyncAPISchema = this.javaClass.getResource("/$asyncapi-$version.json").readText()
+        val asyncAPISchema = this.javaClass.getResource("/$asyncapi-$version.yaml").readText()
         val asyncAPIPSI = psiFileFactory.createFileFromText(
-            "$asyncapi-$version.json",
-                Language.findLanguageByID("JSON")!!,
+                "$asyncapi-$version.yaml",
+                Language.findInstance(YAMLLanguage::class.java),
                 asyncAPISchema
-        ) as JsonFile
+        ) as YAMLFile
 
         collectReferencesToMessages(asyncAPIPSI)
         collectAsyncAPI(asyncAPIPSI, version)
     }
 
-    private fun collectReferencesToMessages(asyncAPI: JsonFile) {
+    private fun collectReferencesToMessages(asyncAPI: YAMLFile) {
         val userSignedUp = listOf(
                 "#/components/messages/UserSignedUp",
                 "#/components/messages/UserSignedUp",
@@ -64,13 +64,13 @@ class JsonFileXPathTest: BasePlatformTestCase() {
         )
         val allReferences = userSignedUp + userSignedOut
 
-        TestCase.assertEquals(allReferences, JsonFileXPath.findText(asyncAPI, "$.channels.*.*.message.\$ref"))
-        TestCase.assertEquals(userSignedUp, JsonFileXPath.findText(asyncAPI, "$.channels.user/signedup.*.message.\$ref"))
-        TestCase.assertEquals(userSignedOut, JsonFileXPath.findText(asyncAPI, "$.channels.user/signedout.*.message.\$ref"))
+        TestCase.assertEquals(allReferences, YamlFileXPath.findText(asyncAPI, "$.channels.*.*.message.\$ref"))
+        TestCase.assertEquals(userSignedUp, YamlFileXPath.findText(asyncAPI, "$.channels.user/signedup.*.message.\$ref"))
+        TestCase.assertEquals(userSignedOut, YamlFileXPath.findText(asyncAPI, "$.channels.user/signedout.*.message.\$ref"))
     }
 
-    private fun collectAsyncAPI(asyncAPI: JsonFile, version: String) {
-        TestCase.assertEquals(listOf(version), JsonFileXPath.findText(asyncAPI, "$.asyncapi"))
+    private fun collectAsyncAPI(asyncAPI: YAMLFile, version: String) {
+        TestCase.assertEquals(listOf(version), YamlFileXPath.findText(asyncAPI, "$.asyncapi"))
     }
 
 }
